Mantis - 9451: Add namespace
Add namespace attribute to entities and entity. The intent is to allow for a large project to break up the namespace and folders that entities are in. i.e. UL.Dss.BusinessLogic.OrderEntry vs UL.Dss.BusinessLogic.Inventory.

Would need to be output in the template, used by the task for the destination folder for the file and for the registration of the entity in the internal types.